+++ NETUA2_Egzaminas/DTOs/PostUserInfoDTO.cs
@@ -4,7 +4,7 @@
 	{
 		public string Name { get; set; }
 		public string Surname { get; set; }
-		public int PersonalID { get; set; }
+		public long PersonalID { get; set; }
 		public int PhoneNumber { get; set; }
 		public string Email { get; set; }
 	}
