+++ NETUA2_Egzaminas/DTOs/UpdateUserInfoDTO.cs
@@ -5,7 +5,7 @@
 		public int Id { get; set; }
 		public string Name { get; set; }
 		public string Surname { get; set; }
-		public int PersonalID { get; set; }
+		public long PersonalID { get; set; }
 		public int PhoneNumber { get; set; }
 		public string Email { get; set; }
 	}
